move trades columns to a separate column
